+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_help.py
@@ -1420,3 +1420,61 @@ helps['identity list-operations'] = """
     type: command
     short-summary: Lists available operations for the Managed Identity provider
 """
+
+helps['sig'] = """
+    type: group
+    short-summary: manage shared image gallery
+"""
+
+helps['sig create'] = """
+    type: command
+    short-summary: create a share image gallery.
+"""
+
+helps['sig list'] = """
+    type: command
+    short-summary: list share image galleries.
+"""
+
+helps['sig update'] = """
+    type: command
+    short-summary: update a share image gallery.
+"""
+
+helps['sig image-definition'] = """
+    type: group
+    short-summary: create an image definition
+"""
+
+helps['sig image-definition create'] = """
+    type: command
+    short-summary: create a gallery image definition
+    examples:
+        - name: Create a linux image defintion
+          text: |
+            az sig image-definition create -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --publisher GreatPublisher --offer GreatOffer --sku GreatSku --os-type linux
+"""
+
+helps['sig image-definition update'] = """
+    type: command
+    short-summary: update a share image defintiion.
+"""
+
+helps['sig image-version'] = """
+    type: group
+    short-summary: create a new version from an image defintion
+"""
+
+helps['sig image-version create'] = """
+    type: command
+    short-summary: creat a new image version
+    examples:
+        - name: add a new version
+          text: |
+            az sig image-version create -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --managed-image /subscriptions/00000000-0000-0000-0000-00000000xxxx/resourceGroups/imageGroups/providers/images/MyManagedImage
+"""
+
+helps['sig image-version update'] = """
+    type: command
+    short-summary: update a share image version
+"""